LazyInstance: Work around issue #21
I think this is a bug in (Ubuntu’s) Python v3.5: In the class LazyInstance,
_instance is a (known) @property. Known attributes are not supposed to be
accessed via the __getattr__ special method. Nonetheless, the attribute
(property) access to self._instance triggers a call to
self.__getattr__('_instance'). Weird. Even weirder that this only happens from
within the ZIP code archive.

I worked around the issue with a regular instance method instead of a property.